After changing name to escriba, I had to use poetry new escriba to recreate the project and directory structure.
Make state an instance in situation, because states can be revisited.
So, when we save data (transformed) in a state, or any state-local data,
then we need separate instances of the state, not just a name.
BUT note: We could also just put such data in the situation,
and thus keep states to be "dumb" (just a name), i.e., static, not instance.
But no, situation is determined during the processing.
So, if you want to have data associated with states, or transforms,
or events, you don't have situations available at design time.
So, this commit has to stand as it is, correct.